fix: Corrigindo valor da proporção que estava calculando errado

+++ src/main/java/br/edu/ifsc/TimetablingGeneticAlgorithm/genetics/Selection.java
@@ -18,7 +18,7 @@ public class Selection {
      * @return Subpopulação de {@link Chromosome}s selecionados.
      * @throws ClassNotFoundException Erro ao tentar obter o pai através do {@code ratingHandler}.
      */
-    public static Chromosome[] rouletteWheel(Chromosome[] chromosomes, int[] ratingHandler, int faA, byte elitismProportion) throws ClassNotFoundException {
+    public static Chromosome[] rouletteWheel(Chromosome[] chromosomes, int[] ratingHandler, int faA, int elitismProportion) throws ClassNotFoundException {
         //O número de cromossomos que serão selecionados é o descontado de quantos já passaram pelo elitismo
         Chromosome[] parents = new Chromosome[(chromosomes.length - elitismProportion)];
 
@@ -60,7 +60,7 @@ public class Selection {
      * @param proportion  O número de {@link Chromosome} que serão selecionados.
      * @return Subpopulação de {@link Chromosome}s selecionados.
      */
-    public static Chromosome[] elitism(Chromosome[] chromosomes, byte proportion) {
+    public static Chromosome[] elitism(Chromosome[] chromosomes, int proportion) {
 
         Chromosome[] chosenChromosomes = new Chromosome[proportion];
         Chromosome[] eliteChromosomes = new Chromosome[chromosomes.length];
@@ -68,7 +68,7 @@ public class Selection {
         //Faz a cópia da população original
         for (int i = 0; i < chromosomes.length; i++) {
             int[] newGenes = new int[chromosomes[i].getGenes().length];
-            System.arraycopy(chromosomes[i].getGenes(),0,newGenes,0,newGenes.length);
+            System.arraycopy(chromosomes[i].getGenes(), 0, newGenes, 0, newGenes.length);
             eliteChromosomes[i] = new Chromosome(newGenes, chromosomes[i].getAvaliation());
         }
